refactor: 공통 컴포넌트 별도 폴더로 분리
