+++ src/components/common/TotalPriceContainer/TotalPriceContainer.style.tsx
